move CollectionCount to new component
Issue: AAH-1677
